LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/parsing/font-stretch-alias-expected.txt


PASS font-stretch is a legacy name alias for font-width
